✨ :: Auth Api 의존성 모듈 추가
