+++ lib/core/config/di/injection.dart
@@ -1,0 +1,8 @@
+import 'package:get_it/get_it.dart';
+import 'package:injectable/injectable.dart';
+import 'package:wisp/core/config/di/injection.config.dart';
+
+final getIt = GetIt.instance;
+
+@injectableInit
+void configureDependencies() => getIt.init();
