+++ lib/core/config/di/network_module.dart
@@ -1,0 +1,8 @@
+import 'package:injectable/injectable.dart';
+import 'package:wisp/core/network/api_service.dart';
+
+@module
+abstract class NetworkModule {
+  @lazySingleton
+  ApiService apiService() => ApiService();
+}
